fixing some lint
