+++ client/src/pages/User.tsx
@@ -1,0 +1,5 @@
+function User() {
+  return <div>User</div>;
+}
+
+export default User;
